Feedback + local notifications fet :D anem avançant :D

+++ lib/Globals/Widgets/GroupOfComponents/Dialogs/FeedbackDialogs/EventFeedbackDialog.dart
@@ -70,7 +70,7 @@ class _EventFeedbackDialogState extends State<EventFeedbackDialog> {
                       color: Theme.of(context).accentColor
                   ),
                   onRatingUpdate: (rating) {
-                    userHasAnsweredFeedback(rating.toInt());
+                    userHasAnsweredFeedback(rating);
                   }
               ),
             ),
@@ -199,7 +199,7 @@ class _EventFeedbackDialogState extends State<EventFeedbackDialog> {
     );
   }
 
-  void userHasAnsweredFeedback(int value) {
+  void userHasAnsweredFeedback(double value) {
     var limitDateToAnswer = eventDate.add(Duration(days: 7));
     print(widget.event.id!);
     if (DateTime.now().isBefore(limitDateToAnswer)) {
